+++ istio/datadog_checks/istio/istio.py
@@ -1,7 +1,9 @@
 # (C) Datadog, Inc. 2018-Present
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
-from datadog_checks.base import ConfigurationError, OpenMetricsBaseCheck
+from six import PY2
+
+from datadog_checks.base import ConfigurationError, OpenMetricsBaseCheck, is_affirmative
 
 from .constants import BLACKLIST_LABELS
 from .legacy_1_4 import LegacyIstioCheck_1_4
@@ -9,6 +11,9 @@ from .metrics import ISTIOD_METRICS
 
 
 class Istio(OpenMetricsBaseCheck):
+    """
+    This is a legacy implementation that will be removed at some point, refer to check.py for the new implementation.
+    """
 
     DEFAULT_METRIC_LIMIT = 0
 
@@ -39,7 +44,20 @@ class Istio(OpenMetricsBaseCheck):
 
     def __new__(cls, name, init_config, instances):
         instance = instances[0]
-        if instance.get('istiod_endpoint'):
-            return super(Istio, cls).__new__(cls)
+
+        if is_affirmative(instance.get('use_openmetrics', False)):
+            if PY2:
+                raise ConfigurationError(
+                    "Openmetrics on this integration is only available when using py3. "
+                    "Check https://docs.datadoghq.com/agent/guide/agent-v6-python-3 "
+                    "for more information"
+                )
+            # TODO: when we drop Python 2 move this import up top
+            from .check import IstioCheckV2
+
+            return IstioCheckV2(name, init_config, instances)
         else:
-            return LegacyIstioCheck_1_4(name, init_config, instances)
+            if instance.get('istiod_endpoint'):
+                return super(Istio, cls).__new__(cls)
+            else:
+                return LegacyIstioCheck_1_4(name, init_config, instances)